requirements.txt
# MCIS - Mysuru Civic Intelligence System
# Python 3.10+

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Vector search & embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# LLM (Ollama client - runs locally)
httpx>=0.26.0

# Config & utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Async support
anyio>=4.2.0
